+++ openapi/impl/java/spring/spring-webapp/src/main/java/uk/co/metagrid/calycopis/offerset/OfferSetRequestParserImpl.java
@@ -763,9 +763,9 @@ log.debug("---- ---- ---- ----");
         log.debug("validate(IvoaSimpleComputeResource)");
         log.debug("Compute [{}])", ((resource != null) ? resource .getName() : "null-resource"));
 
-        Integer MIN_CORES_DEFAULT = 1 ;
-        Integer MAX_CORES_LIMIT   = 16 ;
-        Integer mincores = MIN_CORES_DEFAULT;
+        Long MIN_CORES_DEFAULT =  1L ;
+        Long MAX_CORES_LIMIT   = 16L ;
+        Long mincores = MIN_CORES_DEFAULT;
 
         if (resource.getCores() != null)
             {
